sentiment_bert_finetune-main/backend/requirements.txt
fastapi>=0.104.0,<1.0
uvicorn>=0.24.0,<1.0
transformers>=4.36.0,<5.0
python-dotenv>=1.0.0,<2.0
numpy<2
# PyTorch CPU — install separately if you need GPU support:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
